+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/BackupInstance/Stop-AzDataProtectionBackupInstanceProtection.ps1
@@ -98,7 +98,7 @@ function Stop-AzDataProtectionBackupInstanceProtection
     {
         $parameterSetName = $PsCmdlet.ParameterSetName
         if($parameterSetName -eq "StopViaIdentity"){
-            $Parameter = [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Models.Api20250201.StopProtectionRequest]::new()
+            $Parameter = [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Models.Api20250901.StopProtectionRequest]::new()
 
             $hasResourceGuardOperationRequest = $PSBoundParameters.Remove("ResourceGuardOperationRequest")
             if($hasResourceGuardOperationRequest){